+++ Sources/MapboxMaps/Foundation/MapboxObservable.swift
@@ -3,10 +3,6 @@ import MapboxCoreMaps
 internal protocol MapboxObservableProtocol: AnyObject {
     func subscribe(_ observer: Observer, events: [String])
     func unsubscribe(_ observer: Observer, events: [String])
-    @available(*, deprecated, renamed: "onNext(event:handler:)")
-    func onNext(_ eventTypes: [MapEvents.EventKind], handler: @escaping (Event) -> Void) -> Cancelable
-    @available(*, deprecated, renamed: "onEvery(event:handler:)")
-    func onEvery(_ eventTypes: [MapEvents.EventKind], handler: @escaping (Event) -> Void) -> Cancelable
     func onNext<Payload>(event: MapEvents.Event<Payload>, handler: @escaping (MapEvent<Payload>) -> Void) -> Cancelable
     func onEvery<Payload>(event: MapEvents.Event<Payload>, handler: @escaping (MapEvent<Payload>) -> Void) -> Cancelable
     func performWithoutNotifying(_ block: () -> Void)
@@ -76,11 +72,6 @@ internal final class MapboxObservable: MapboxObservableProtocol {
         }
     }
 
-    @available(*, deprecated, renamed: "onNext(event:handler:)")
-    internal func onNext(_ eventTypes: [MapEvents.EventKind], handler: @escaping (Event) -> Void) -> Cancelable {
-        return onNext(eventTypes.map(\.rawValue), handler: handler)
-    }
-
     internal func onNext<Payload>(event: MapEvents.Event<Payload>, handler: @escaping (MapEvent<Payload>) -> Void) -> Cancelable {
         return onNext([event.name]) { event in
             handler(MapEvent(event: event))
@@ -105,11 +96,6 @@ internal final class MapboxObservable: MapboxObservableProtocol {
             }
         })
         return cancelable
-    }
-
-    @available(*, deprecated, renamed: "onEvery(event:handler:)")
-    internal func onEvery(_ eventTypes: [MapEvents.EventKind], handler: @escaping (Event) -> Void) -> Cancelable {
-        return onEvery(eventTypes.map(\.rawValue), handler: handler)
     }
 
     internal func onEvery<Payload>(event: MapEvents.Event<Payload>, handler: @escaping (MapEvent<Payload>) -> Void) -> Cancelable {